Filtering + inventory + log + servers
Created regex filtering for any special characters in drink name. If there is a space I changed it to an underscore for simplicity. I also added an inventory dictionary to keep track of supplies as well as a log for each server and what they order. I created a simple server dictionary with our names in order to make it easier to choose a server. I still can't run the program though so I do not know what it looks like right now or if what I've done even works currently lol.